+++ common/src/python/configs/ingest_configs.py
@@ -1,11 +1,19 @@
 """Form ingest configurations."""
 
+from json.decoder import JSONDecodeError
 from pathlib import Path
 from string import Template
 from typing import Any, Dict, List, Literal, Optional
 
+from gear_execution.gear_trigger import GearInfo
 from keys.keys import DefaultValues
-from pydantic import BaseModel, Field, RootModel
+from pydantic import BaseModel, Field, RootModel, ValidationError
+
+PipelineType = Literal['submission', 'finalization']
+
+
+class ConfigsError(Exception):
+    pass
 
 
 class LabelTemplate(BaseModel):
@@ -124,3 +132,60 @@ class FormProjectConfigs(BaseModel):
     accepted_modules: List[str]
     legacy_project_label: Optional[str] = DefaultValues.LEGACY_PRJ_LABEL
     module_configs: Dict[str, ModuleConfigs]
+
+    def get_module_dependencies(self, module: str) -> Optional[List[str]]:
+        """Returns the list of dependent modules for a given module.
+
+        Args:
+            module: module label
+
+        Returns:
+            List[str](optional): list of dependent module labels if found
+        """
+
+        dependent_modules = []
+        for module_label, config in self.module_configs.items():
+            if (config.supplement_module
+                    and config.supplement_module.exact_match
+                    and config.supplement_module.label == module.upper()):
+                dependent_modules.append(module_label)
+
+        return dependent_modules
+
+
+class Pipeline(BaseModel):
+    """Defines model for form scheduler pipeline."""
+    name: PipelineType
+    modules: List[str]
+    tags: List[str]
+    extensions: List[str]
+    starting_gear: GearInfo
+    notify_user: bool = False
+
+
+class PipelineConfigs(BaseModel):
+    gears: List[str]
+    pipelines: List[Pipeline]
+
+    @classmethod
+    def load_form_pipeline_configurations(
+            cls, config_file_path: str) -> 'PipelineConfigs':
+        """Load the form pipeline configs from the pipeline configs file.
+
+        Args:
+            config_file_path: the form module configs file path
+
+        Returns:
+            PipelineConfigs
+
+        Raises:
+            ConfigsError if failed to load the configs file
+        """
+
+        try:
+            with open(config_file_path, mode='r',
+                      encoding='utf-8-sig') as configs_file:
+                return PipelineConfigs.model_validate_json(configs_file.read())
+        except (FileNotFoundError, JSONDecodeError, TypeError,
+                ValidationError) as error:
+            raise ConfigsError(error) from error