+++ docs/form_scheduler/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.1.0
+* Adds functionality for handling multiple pipelines (submission, finalization)
+  
 ## 1.0.5
 * Updates to process all files in a given queue before moving to next queue/module